requirements.txt
torch==2.2.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
transformers>=4.44.0
nltk>=3.8.1
textblob>=0.17.1
emoji>=2.5.0
pyspellchecker>=0.7.5
tqdm>=4.66.0
regex>=2023.6.3
datasets>=2.15.1
sentencepiece>=0.1.99
protobuf>=4.24.3
